+++ ipns/errors.go
@@ -47,8 +47,8 @@ var ErrInvalidName = errors.New("name is invalid")
 // ErrInvalidPath is returned when an IPNS [Record] has an invalid path.
 var ErrInvalidPath = errors.New("value is not a valid content path")
 
-// ErrMetadataNotFound is returned when when IPNS metadata dictionary key does not exist
+// ErrMetadataNotFound is returned when an IPNS metadata dictionary key does not exist.
 var ErrMetadataNotFound = errors.New("metadata key not found in record")
 
-// ErrMetadataConflict is returned when metadata key uses a reserved keyword
+// ErrMetadataConflict is returned when a metadata key uses a reserved IPNS field name.
 var ErrMetadataConflict = errors.New("metadata key uses reserved name")